development: Fixed test script
Fixed use case where the test script is the first to look for the
directories.  It now switches on whether or not they exist at test-time.
How did increment stop working?!
Changes to be committed:
	modified:   devops/scripts/filename_test.sh

+++ devops/scripts/filename_test.sh
@@ -1,14 +1,32 @@
 make filename_test
 make source08
 
-rm -f /tmp/watch/*
-rm -f /tmp/watch/processed/*
+WATCH_DIR=/tmp/watch/
+PROCESS_DIR=/tmp/watch/processed/
 TEST_NUM=1
-while [[ `ls /tmp/watch/processed | wc -l` -eq 0 ]] && [[ `ls /tmp/watch | wc -l` -eq 1 ]]
+
+# SETUP ENVIRONMENT
+# Watch
+if [ -d $WATCH_DIR ]
+then
+    rm -f /tmp/watch/*
+else
+	mkdir $WATCH_DIR
+fi
+# Processed
+if [ -d $PROCESS_DIR ]
+then
+    rm -f /tmp/watch/processed/*
+else
+	mkdir $PROCESS_DIR
+fi
+
+# TEST CODE
+while [ `ls /tmp/watch/processed | wc -l` -eq 0 ] && [ `ls /tmp/watch | wc -l` -eq 1 ]
 do
 	echo "Test: $TEST_NUM"
 	echo -n "some_filename.txt" | radamsa > filename_test.txt && sudo $1 filename_test.txt
-	((TEST_NUM=TEST_NUM+1))
+	TEST_NUM=$((TEST_NUM+1))
 done
 ls /tmp/watch/
 ls /tmp/watch/processed/
